// src/interfaces/agency.interface.ts
export default interface IAgency extends Document {
    agencyName: string;
    accountNumber: number
    companyName: string;
    doingBusinessAs: string;
    taxId: string;
    phone: string;
    fax: string;
    address: string;
    streetAddress: string;
    city: string;
    mailingAddressStateCode: string;
    state: string;
    mailingAddressCountry: string;
    postalCode: string;
    contactFirstName: string;
    contactLastName: string;
    // contactEmail: string;
    contactMobileNumber: string;
    contactOfficePhone: string;
    accreditation: string;
    agencyAdminEmail: string;
    agencyAdminFirstName: string;
    agencyAdminLastName: string;
    agencyAdminMobile: string;
    agencyOperationalStatus: string;
    agencySubmissionStatus: string;
    billingAddressCity: string;
    billingAddressLine1: string;
    billingAddressLine2: string;
    billingAddressStateName: string;
    billingAddressZipCode: string;
    companyType: string;
    email: string;
    medicareProviderNo: string;
    npi: string;
    stateHhaLicenseNo: string;
    submitterId: string;
    subscriptionPlanId: string;
    agencyShortCode: string;
    billingAddressAreaCode: string;
    hiqAdminComment: string;
    comments: string;
    baaEnvelopeId: string;
    baaEnvelopeSentDate: string;
    baaEnvelopeStatus: string;
    discountId: string;
    website: string;
    demoRequestId: string;  
    contactPerson1Branch: string;
    contactPerson1Email: string;
    contactPerson1Name: string;
    contactPerson1Phone: string;
    contactPerson1Title: string;
    contactPerson2Branch: string;
    contactPerson2Email: string;
    contactPerson2Name: string;
    contactPerson2Phone: string;
    contactPerson2Title: string;
    contactPerson3Branch: string;
    contactPerson3Email: string;
    contactPerson3Name: string;
    contactPerson3Phone: string;
    contactPerson3Title: string;
    agencyId: string;
    status: string;
    isDeleted: boolean;
    isAproved: boolean
    isForProfit: boolean;
    isActive: boolean;
}